Revert "Move plugin message handling to plugin_messages::PluginMessageHandler"
This reverts commit f91f592f73a7b36b411332e22492a0543eafc740.
