+++ src/server/plugin_messages.rs
@@ -3,47 +3,6 @@ use crate::protocol::Serializable;
 use crate::protocol::packet::play::serverbound::PluginMessageServerbound;
 use crate::protocol::packet::play::serverbound::PluginMessageServerbound_i16;
 
-pub struct PluginMessageHandler {
-}
-
-impl PluginMessageHandler {
-    pub fn on_plugin_message_clientbound(&mut self, channel: &str, data: &[u8]) {
-        println!("Received plugin message: channel={}, data={:?}", channel, data);
-
-        match channel {
-            // TODO: "REGISTER" => 
-            // TODO: "UNREGISTER" =>
-            "FML|HS" => {
-                // https://wiki.vg/Minecraft_Forge_Handshake
-                let discriminator = data[0];
-
-                match discriminator {
-                    0 => {
-                        // ServerHello
-                        let fml_protocol_version = data[1];
-                        let dimension = if fml_protocol_version > 1 {
-                            use byteorder::{BigEndian, ReadBytesExt};
-                            let dimension = (&data[2..2 + 4]).read_u32::<BigEndian>().unwrap();
-                            Some(dimension)
-                        } else {
-                            None
-                        };
-
-                        println!("FML|HS ServerHello: fml_protocol_version={}, dimension={:?}", fml_protocol_version, dimension);
-
-                        // TODO: send reply
-                    },
-                    _ => {
-                        println!("Unhandled FML|HS packet: discriminator={}", discriminator);
-                    }
-                }
-            }
-            _ => ()
-        }
-    }
-}
-
-
 pub struct Brand {
     pub brand: String,
 }
